+++ src/features/create-post/api/createPostApi.ts
@@ -1,7 +1,7 @@
 import { axiosInstance } from '@/shared/api/base/axiosInstance';
-import type { CreatePostInput } from '../model/validation';
+import type { CreatePostFormValues } from '../lib/post.scheme';
 
-export async function createPost(input: CreatePostInput) {
+export async function createPost(input: CreatePostFormValues) {
   const { data } = await axiosInstance.post('/api/posts', {
     title: input.title,
     content: input.content,
