+++ src/features/create-post/model/useCreatePost.ts
@@ -1,9 +1,9 @@
 import { useMutation } from '@tanstack/react-query';
 import { createPost } from '../api/createPostApi';
-import type { CreatePostInput } from './validation';
+import type { CreatePostFormValues } from '../lib/post.scheme';
 
 export function useCreatePost() {
   return useMutation({
-    mutationFn: (input: CreatePostInput) => createPost(input),
+    mutationFn: (input: CreatePostFormValues) => createPost(input),
   });
 }
